+++ pyproject.toml
@@ -32,6 +32,7 @@ mypy = "^1.11.2"
 types-passlib = "^1.7.7.20240819"
 pydantic = "^2.8.2"
 pytest-mock = "^3.14.0"
+testcontainers = {extras = ["mysql"], version = "^4.8.0"}
 
 
 [build-system]
